+++ tests/schemas.py
@@ -1,7 +1,9 @@
 from marshmallow import EXCLUDE
 
 from apiflask import Schema
+from apiflask.fields import File
 from apiflask.fields import Integer
+from apiflask.fields import List
 from apiflask.fields import String
 
 
@@ -31,6 +33,23 @@ class QuerySchema(Schema):
         unknown = EXCLUDE
 
     id = Integer(load_default=1)
+
+
+class FormSchema(Schema):
+    name = String()
+
+
+class FilesSchema(Schema):
+    image = File()
+
+
+class FilesListSchema(Schema):
+    images = List(File())
+
+
+class FormAndFilesSchema(Schema):
+    name = String()
+    image = File()
 
 
 class PaginationSchema(Schema):